1/12

+++ app/Views/rekap/rekap_dash.php
@@ -21,21 +21,18 @@
                     <!-- </div> -->
 
                     <div class="card-body">
-                        
+
                         <table id="resp_table" class="display table table-bordered table-hover">
                             <thead>
                                 <tr title="Klik untuk mengurutkan data">
                                     <!-- <th>ID Penjualan</th> -->
                                     <th>Nama Alias</th>
-
                                     <th>Rentang Tanggal Awal</th>
                                     <th>Rentang Tanggal Akhir</th>
                                     <th>Tanggal dan Waktu Memasukkan Data</th>
                                     <!-- <th>Nilai Davies-Bouldin Index</th>
                                     <th>Selisih Simpangan</th> -->
                                     <th>Aksi</th>
-
-
                                 </tr>
                             </thead>
                             <tbody>
@@ -45,9 +42,10 @@
                                 foreach ($data_penjualan as $dp) {
                                 ?>
                                     <tr>
-                                        <!-- <td> <?//= $dp['penjualan_id']; ?> </td> -->
+                                        <!-- <td> <? //= $dp['penjualan_id']; 
+                                                    ?> </td> -->
                                         <td>
-                                            <?= $dp['nama_alias']; 
+                                            <?= $dp['nama_alias'];
                                             ?>
                                         </td>
                                         <td>
@@ -62,7 +60,8 @@
 
                                         </td>
                                         <td>
-                                            <?//= $dp['timestamp_enterdata']; ?>
+                                            <? //= $dp['timestamp_enterdata']; 
+                                            ?>
                                             <?php echo date('d-m-Y - H:i:s', strtotime($dp['timestamp_enterdata'])); ?>
 
                                         </td>
@@ -70,26 +69,27 @@
                                         <!-- <td> <?//= $dp['selisih_simpangan']; ?> </td> -->
 
                                         <td>
-                                            <a href="#">
-                                                <a href="<?= base_url(); ?>/rekap_data/rekap_tr/<?= $dp['penjualan_id']; ?>">
-                                                    <!-- <button type="button" class="btn btn-primary"> -->
-                                                    <button type="button" class="btn btn-primary" data-toggle="tooltip" data-placement="top" title="Detail Data Rekap">
-                                                        <i class="fa fa-eye" aria-hidden="true"></i>
-                                                    </button>
-                                                    
-                                                </a>
+                                            <a href="<?= base_url(); ?>/rekap_data/rekap_tr/<?= $dp['penjualan_id']; ?>">
+                                                <button type="button" class="btn btn-primary" data-toggle="tooltip" data-placement="top" title="Detail Data Rekap">
+                                                    <i class="fa fa-eye" aria-hidden="true"></i>
+                                                </button>
                                             </a>
-                                            <a href="#">
-                                                <!-- <a href="<?= base_url(); ?>/rekap_data/delete_rekap/<?= $dp['penjualan_id']; ?>"> -->
-                                                <a href="<?= base_url(); ?>/rekap_data/delete_rekap/<?= $dp['penjualan_id']; ?>
-                                                " onclick="return confirm('Yakin ingin menghapus data?');">
-                                                    <!-- <button type="button" class="btn btn-danger"> -->
-                                                    <button type="button" class="btn btn-danger" data-toggle="tooltip" data-placement="top" title="Hapus Data Rekap">
 
-                                                        <i class="fa fa-trash" aria-hidden="true"></i>
-                                                    </button>
-                                                </a>
+                                            <!-- <a href="#"></a>  -->
+                                            <!-- <a href="<?//= base_url(); ?>/rekap_data/editAlias/<?//= $dp['penjualan_id']; ?>"> -->
+                                            <button type="button" class="btn btn-success" data-toggle="modal" data-target="#modal-default" 
+                                            data-placement="top" title="Edit Nama Alias Data Rekapitulasi">
+                                                <i class="fa fa-edit" aria-hidden="true"></i>
+                                            </button>
                                             </a>
+
+                                            <a href="<?= base_url(); ?>/rekap_data/delete_rekap/<?= $dp['penjualan_id']; ?>" onclick="return confirm('Yakin ingin menghapus data?');">
+                                                <button type="button" class="btn btn-danger" data-toggle="tooltip" data-placement="top" title="Hapus Data Rekap">
+                                                    <i class="fa fa-trash" aria-hidden="true"></i>
+                                                </button>
+                                            </a>
+
+
                                         </td>
 
                                     </tr>
@@ -100,6 +100,43 @@
 
 
                         </table>
+
+                        <!-- modal edit -->
+                        <div class="modal fade" id="modal-default">
+                            <form method="post" enctype="multipart/form-data" action="<?= base_url(); ?>/rekap_data/editAlias/<?= $dp['penjualan_id']; ?>">
+                                <div class="modal-dialog">
+                                    <div class="modal-content">
+                                        <div class="modal-header">
+                                            <h4 class="modal-title">Edit Nama Alias Data Rekapitulasi</h4>
+                                            <button type="button" class="close" data-dismiss="modal" aria-label="Close">
+                                                <span aria-hidden="true">&times;</span>
+                                            </button>
+                                        </div>
+                                        <div class="modal-body">
+                                            
+                                            <div class="card-body ">
+
+                                                <div class="form-group">
+                                                    <label for="alias">Nama Baru</label>
+                                                    <input class="form-control" type="text" 
+                                                    placeholder="Nama Alias File Baru" name="nama_alias" required>
+                                                </div>
+                                                
+                                            </div>
+                                            <!-- /.card-body -->
+                                        </div>
+                                        <div class="modal-footer justify-content-between">
+                                            <button type="button" class="btn btn-default" data-dismiss="modal">Batal</button>
+                                            <button type="submit" class="btn btn-success">Simpan Nama Baru</button>
+                                        </div>
+                                    </div>
+                                    <!-- /.modal-content -->
+                                </div>
+                                <!-- /.modal-dialog -->
+                            </form>
+                        </div>
+
+
                     </div>
 
                 </div>
